MGMT-832 Add deploy paramter to oc-deploy target. DEPLOY_NAMESPACE - true by default. DELETE_NAMESPACE - true by default, used in clear-deployment target. INGRESS_DOMAIN - In case of TARGET=oc-ingress the script will support specific domain or getting a default domain from ingresscontrollers.operator.openshift.io default openshift ingress operator.
tools/utils.py cleanup

+++ tools/clear_deployment.py
@@ -1,8 +1,15 @@
 import utils
+import argparse
+
+parser = argparse.ArgumentParser()
+parser.add_argument("--delete-namespace", type=lambda x: (str(x).lower() == 'true'), default=True)
+args = parser.parse_args()
+
 
 def main():
     print(utils.check_output("kubectl delete all --all -n assisted-installer 1> /dev/null ; true"))
-    print(utils.check_output("kubectl delete namespace assisted-installer 1> /dev/null ; true"))
+    if args.delete_namespace is True:
+        print(utils.check_output("kubectl delete namespace assisted-installer 1> /dev/null ; true"))
 
 if __name__ == "__main__":
     main()
